jwt: user, follow, blog and like module

+++ resource/blog.py
@@ -1,4 +1,5 @@
 from flask_restful import Resource, reqparse
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from datetime import datetime
 from model.blog import BlogModel
 from model.user import UserModel
@@ -9,15 +10,15 @@ class BlogsResource(Resource):
     def get(self):
         return [blog.json() for blog in BlogModel.find_all()]
 
+    @jwt_required()
     def post(self):
         _parser = reqparse.RequestParser()
         _parser.add_argument('title', type=str, required=True)
         _parser.add_argument('content', type=str, required=True)
-        _parser.add_argument('username', type=str, required=True)
         payload = _parser.parse_args()
         title = payload.get('title')
         content = payload.get('content')
-        username = payload.get('username')
+        username = get_jwt_identity()
         post_time = datetime.now()
         if BlogModel.find_by_title(title):
             return {'message': f'similar title exist', 'title': title}, 400
@@ -41,6 +42,7 @@ class BlogResource(Resource):
 
         return decorate()
 
+    @jwt_required()
     def put(self, blog_id):
         _parse = reqparse.RequestParser()
         _parse.add_argument('title', type=str, required=True)
@@ -50,8 +52,11 @@ class BlogResource(Resource):
         content = payload.get('content')
 
         @get_one_or_404(BlogModel, id=blog_id)
+        @get_one_or_404(UserModel, username=get_jwt_identity())
         def decorate(*args):
-            (blog,) = args
+            (blog, user) = args
+            if blog.post_by != user:
+                return {'message': 'this is not your blog'}, 403
             blog.title = title
             blog.content = content
             blog.save()
@@ -59,10 +64,14 @@ class BlogResource(Resource):
 
         return decorate()
 
+    @jwt_required()
     def delete(self, blog_id):
         @get_one_or_404(BlogModel, id=blog_id)
+        @get_one_or_404(UserModel, username=get_jwt_identity())
         def decorate(*args):
-            (blog,) = args
+            (blog, user) = args
+            if blog.post_by != user:
+                return {'message': 'this is not your blog'}, 403
             response = blog.json()
             blog.delete()
             return {'message': f'blog deleted', 'object': response}
